backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pyjwt==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
reportlab==4.1.0
python-dotenv==1.0.0
eventlet==0.35.2
python-socketio==5.11.0
obd==0.7.1
requests==2.31.0
gunicorn==21.2.0
